Move AssemblyCompilerCallbacks to separate file
