+++ Cql/CodeGeneration.NET/AssemblyCompiler.cs
@@ -292,17 +292,6 @@ namespace Hl7.Cql.CodeGeneration.NET
         }
     }
 
-    internal class AssemblyCompilerCallbacks
-    {
-        private readonly Action<(string name, Stream stream, bool isTuple)>? _onBeforeCompileStream;
-
-        public AssemblyCompilerCallbacks(Action<(string name, Stream stream, bool isTuple)>? onBeforeCompileStream = null) => 
-            _onBeforeCompileStream = onBeforeCompileStream;
-
-        public void OnBeforeCompileStream(string name, Stream stream, bool isTuple) => 
-            _onBeforeCompileStream?.Invoke((name, stream, isTuple));
-    }
-
     internal readonly record struct ExpectedTuplesBeforeLibrariesError()  : ICqlError
     {
         public string GetMessage() => "Expected all tuple types before library types";
